stages/train/requirements.txt
scikit-learn==1.4.0
pandas==2.2.0
pyyaml==6.0.1
mlflow==2.11.0
